KAFKA-14588 [2/N] ConfigCommandTest rewritten in java
